Updating Displacement computation
Last commit of this file was not working : we did not store the vectorfields at intended (because a vectorfield object, being a vectorarray, does not store the data but a pointer on the data). We therefore allocate memory for two vectorfields at a time that we will be able to substract to get the displacement map. The displacement computation now works and can be used as a stopping condition.