ENH : modifying the reader and writer to get 3D vector iamges instead of 4D scalar images (in 3D)
* LDDMM, writing transformation map :
writeFileITK( ) has been edited, in case the voxel-stored data dimension is the same as the image dimension, to write a DeformationImageType [1] instead of the CalaTK ITKVectorImage object [2].
For example in 3D, it was writing a 4D scalar object instead of a 3D vector object.
To give to the writer the good parameter, we had to modify convertToITK( ) to return a DeformationImageType from the VectorImageType given.
=> ISSUE : Even though unu head on the written file or print() the DeformationImageType seem to give us the new structure we were looking for, we can not visualize this displacement field or use it with our current tools.

* applyMap, reading transformation map and writing computed image :
We should not have to modify writeFileITK( ) anymore since it was making the distinction between scalar image and vector image ( line 2473/2502 VectorImageUtils.txx )
The only thing we had to do was therefore change the reader to make sure it would be able to convert the DeformationImageType written with LDDMM into this VectorImageType in CalaTK.
readFileITK( ) was therefore modified to read the file as a DeformationImageType and not a ITKVectorImage.
After this, we had to modify convertFromITK( ) so that it would create this VectorImageType from a DeformationImageType and not ITKVectorImage. I used the inverse method used in convertToITK.
=> ISSUE : output file is written as a 3D vector image instead of a 3D scalar image. This should not happen even if we know that there might be an issue with the transformation map we are reading.

[1] DeformationImageType
typedef itk::Vector<T,VImageDimension> DeformationPixelType;
typedef itk::Image<DeformationPixelType, VImageDimension> DeformationImageType;

[2] ITKVectorImage
struct ITKVectorImage { typedef itk::Image< T, VImageDimension + 1 > Type; };